Move new tessellator tests into the fill_test module.
